Fix click handlers on PaginationDots
The handleDotClick method used by PaginationDots component was trying
to calculate mutate the `SliderState` directly, using a dispatch call.
However, the value for the state variable `currentSlide` was wrongly calculated,
which meant the SliderState was not update as expected.

This incongruence caused slides to not be rendered when users clicked in
a PaginationDot to navigate, since the `useSliderVisibility` would
be fooled into returning `false` for the `isItemVisible` variable.